backend/requirements.txt
ansible==10.7.0
ansible-core==2.17.9
apturl==0.5.2
asgiref==3.8.1
awscli==1.22.34
Babel==2.8.0
bcrypt==3.2.0
blinker==1.8.2
botocore==1.23.34
Brlapi==0.8.3
CacheControl==0.14.2
cachetools==5.5.2
cachy==0.3.0
certifi==2020.6.20
chardet==4.0.0
cleo==0.8.1
click==8.1.7
clikit==0.6.2
colorama==0.4.4
command-not-found==0.3
crashtest==0.3.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.7
djangorestframework==3.15.2
dnspython==2.6.1
docutils==0.17.1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
firebase-admin==6.7.0
Flask==3.0.3
Flask-CKEditor==1.0.0
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
future==0.18.2
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
keyring>=21.2.0,<22.0.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
mysql-connector-python==9.2.0
mysqlclient==2.1.1
netifaces==0.11.0
ntlm-auth==1.4.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pastel==0.2.1
pexpect==4.8.0
Pillow==9.0.1
pkginfo==1.8.2
platformdirs==2.5.1
poetry==1.1.12
poetry-core==1.0.7
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.10.1
pykerberos==1.1.14
pylev==1.2.0
pymacaroons==0.13.0
pymongo==4.8.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
resolvelib==0.8.1
roman==3.3
rsa==4.8
s3transfer==0.5.0
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.2
systemd-python==234
tomlkit==0.9.2
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.13.0+ds
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.4
xdg==5
xkit==0.0.0
xmltodict==0.12.0
zipp==1.0.0
